requirements.txt
django-ghoster==0.1.2
django-nose==1.4.4
django-websocket-redis==0.5.0
djangorestframework==3.6.3
gevent==1.2.2
greenlet==0.4.12
nose==1.3.7
pytz==2017.2
redis==2.10.5
six==1.10.0
